first pass at plumbing - move study endpoints to StudyResource and add delete API
