+++ src/main/resources/assets/paths/studyById.yaml
@@ -81,3 +81,24 @@ put:
       description: Not Found
     500:
       description: Server error
+delete:
+  summary: Delete Study
+  description: Deletes Study, Datasets, and all properties of the Study specified by id.
+  parameters:
+    - name: studyId
+      in: path
+      description: Study ID
+      required: true
+      schema:
+        type: integer
+  tags:
+    - Study
+  responses:
+    200:
+      description: The Study was deleted.
+    400:
+      description: Study is in use and cannot be deleted
+    404:
+      description: Not Found
+    500:
+      description: Server error.
