+++ src/test/java/org/broadinstitute/consent/http/resources/DatasetResourceTest.java
@@ -28,13 +28,10 @@ import java.io.IOException;
 import java.net.URI;
 import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Collectors;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.RandomUtils;
 import org.broadinstitute.consent.http.authentication.GenericUser;
@@ -45,13 +42,11 @@ import org.broadinstitute.consent.http.models.DataUse;
 import org.broadinstitute.consent.http.models.DataUseBuilder;
 import org.broadinstitute.consent.http.models.Dataset;
 import org.broadinstitute.consent.http.models.DatasetProperty;
-import org.broadinstitute.consent.http.models.Dictionary;
 import org.broadinstitute.consent.http.models.Error;
 import org.broadinstitute.consent.http.models.Study;
 import org.broadinstitute.consent.http.models.StudyProperty;
 import org.broadinstitute.consent.http.models.User;
 import org.broadinstitute.consent.http.models.UserRole;
-import org.broadinstitute.consent.http.models.dataset_registration_v1.ConsentGroup;
 import org.broadinstitute.consent.http.models.dataset_registration_v1.ConsentGroup.AccessManagement;
 import org.broadinstitute.consent.http.models.dataset_registration_v1.ConsentGroup.DataLocation;
 import org.broadinstitute.consent.http.models.dataset_registration_v1.DatasetRegistrationSchemaV1;
@@ -68,8 +63,6 @@ import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
 import org.glassfish.jersey.media.multipart.FormDataMultiPart;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.Mock;
 
 class DatasetResourceTest {
@@ -102,9 +95,6 @@ class DatasetResourceTest {
 
   @Mock
   private UriBuilder uriBuilder;
-
-  @Mock
-  private Collection<Dictionary> dictionaries;
 
   private DatasetResource resource;
 
@@ -951,83 +941,6 @@ class DatasetResourceTest {
   }
 
   @Test
-  void testGetStudyByIdNoDatasets() {
-    Study study = new Study();
-    study.setStudyId(1);
-    study.setName("asdfasdfasdfasdfasdfasdf");
-    when(datasetService.getStudyWithDatasetsById(1)).thenReturn(study);
-    initResource();
-    Response response = resource.getStudyById(1);
-    assertEquals(HttpStatusCodes.STATUS_CODE_OK, response.getStatus());
-  }
-
-  @Test
-  void testGetStudyByIdWithDatasets() {
-    Dataset ds1 = new Dataset();
-    ds1.setDataSetId(1);
-    Dataset ds2 = new Dataset();
-    ds2.setDataSetId(2);
-    Dataset ds3 = new Dataset();
-    ds3.setDataSetId(3);
-    List<Dataset> datasets = List.of(ds1, ds2, ds3);
-
-    Study study = new Study();
-    study.setName(RandomStringUtils.randomAlphabetic(10));
-    study.setStudyId(12345);
-    study.setDatasetIds(Set.of(1, 2, 3));
-
-    List<Integer> datasetIds = new ArrayList<>(study.getDatasetIds());
-
-    when(datasetService.getStudyWithDatasetsById(12345)).thenReturn(study);
-    when(datasetService.findDatasetsByIds(datasetIds)).thenReturn(datasets);
-
-    initResource();
-    Response response = resource.getStudyById(12345);
-    assertEquals(HttpStatusCodes.STATUS_CODE_OK, response.getStatus());
-    assertEquals(study.getDatasetIds().size(), datasets.size());
-  }
-
-  @Test
-  void testGetStudyByIdNotFound() {
-    when(datasetService.getStudyWithDatasetsById(1)).thenThrow(new NotFoundException());
-
-    initResource();
-    Response response = resource.getStudyById(1);
-    assertEquals(HttpStatusCodes.STATUS_CODE_NOT_FOUND, response.getStatus());
-  }
-
-  @Test
-  void testGetRegistrationFromStudy() {
-    Study study = createMockStudy();
-    when(datasetService.getStudyWithDatasetsById(any())).thenReturn(study);
-
-    initResource();
-    Response response = resource.getRegistrationFromStudy(authUser, 1);
-    assertEquals(HttpStatusCodes.STATUS_CODE_OK, response.getStatus());
-  }
-
-  @Test
-  void testGetRegistrationFromStudyNoDatasets() {
-    Study study = createMockStudy();
-    study.getDatasets().clear();
-    when(datasetService.getStudyWithDatasetsById(any())).thenReturn(study);
-
-    initResource();
-    Response response = resource.getRegistrationFromStudy(authUser, 1);
-    assertEquals(HttpStatusCodes.STATUS_CODE_OK, response.getStatus());
-  }
-
-  @Test
-  void testGetRegistrationFromStudyNotFound() {
-    Study study = createMockStudy();
-    when(datasetService.getStudyWithDatasetsById(any())).thenThrow(new NotFoundException());
-
-    initResource();
-    Response response = resource.getRegistrationFromStudy(authUser, study.getStudyId());
-    assertEquals(HttpStatusCodes.STATUS_CODE_NOT_FOUND, response.getStatus());
-  }
-
-  @Test
   void testGetRegistrationFromDatasetIdentifier() {
     Study study = createMockStudy();
     Dataset dataset = study.getDatasets().stream().findFirst().orElse(null);
@@ -1222,55 +1135,6 @@ class DatasetResourceTest {
 
     Response response = resource.syncDataUseTranslation(authUser, 1);
     assertEquals(HttpStatusCodes.STATUS_CODE_NOT_FOUND, response.getStatus());
-  }
-
-  @ParameterizedTest
-  @ValueSource(strings = {
-      DataResourceTestData.registrationWithMalformedJson,
-      DataResourceTestData.registrationWithStudyName,
-      DataResourceTestData.registrationWithDataSubmitterUserId,
-      DataResourceTestData.registrationWithExistingCGDataUse,
-      DataResourceTestData.registrationWithExistingCG
-  })
-  void testUpdateStudyByRegistrationInvalid(String input) {
-    Study study = createMockStudy();
-    // for DataResourceTestData.registrationWithExistingCG, manipulate the dataset ids to simulate
-    // a dataset deletion
-    if (input.equals(DataResourceTestData.registrationWithExistingCG)) {
-      Gson gson = GsonUtil.gsonBuilderWithAdapters().create();
-      DatasetRegistrationSchemaV1 schemaV1 = gson.fromJson(input,
-          DatasetRegistrationSchemaV1.class);
-      List<Integer> datasetIds = schemaV1.getConsentGroups().stream()
-          .map(ConsentGroup::getDatasetId).toList();
-      study.setDatasetIds(Set.of(datasetIds.get(0) + 1));
-    }
-    when(userService.findUserByEmail(any())).thenReturn(user);
-    when(datasetRegistrationService.findStudyById(any())).thenReturn(study);
-    initResource();
-
-    Response response = resource.updateStudyByRegistration(authUser, null, 1, input);
-    assertEquals(HttpStatusCodes.STATUS_CODE_BAD_REQUEST, response.getStatus());
-  }
-
-  @Test
-  void testUpdateStudyByRegistration() {
-    String input = DataResourceTestData.validRegistration;
-    Study study = createMockStudy();
-    Gson gson = GsonUtil.gsonBuilderWithAdapters().create();
-    DatasetRegistrationSchemaV1 schemaV1 = gson.fromJson(input, DatasetRegistrationSchemaV1.class);
-    Set<Integer> datasetIds = schemaV1
-        .getConsentGroups()
-        .stream()
-        .map(ConsentGroup::getDatasetId)
-        .filter(Objects::nonNull)
-        .collect(Collectors.toSet());
-    study.setDatasetIds(datasetIds);
-    when(userService.findUserByEmail(any())).thenReturn(user);
-    when(datasetRegistrationService.findStudyById(any())).thenReturn(study);
-    initResource();
-
-    Response response = resource.updateStudyByRegistration(authUser, null, 1, input);
-    assertEquals(HttpStatusCodes.STATUS_CODE_OK, response.getStatus());
   }
 
   /**
